+++ packages/astro/src/core/compile/compile.ts
@@ -3,7 +3,7 @@ import type { ResolvedConfig } from 'vite';
 import type { AstroConfig } from '../../@types/astro';
 
 import { transform } from '@astrojs/compiler';
-import { fileURLToPath } from 'url';
+import { fileURLToPath } from 'node:url';
 import { normalizePath } from 'vite';
 import { AggregateError, AstroError, CompilerError } from '../errors/errors.js';
 import { AstroErrorData } from '../errors/index.js';
@@ -45,6 +45,8 @@ export async function compile({
 			astroGlobalArgs: JSON.stringify(astroConfig.site),
 			scopedStyleStrategy: astroConfig.scopedStyleStrategy,
 			resultScopedSlot: true,
+			experimentalTransitions: astroConfig.experimental.viewTransitions,
+			transitionsAnimationURL: 'astro/components/viewtransitions.css',
 			preprocessStyle: createStylePreprocessor({
 				filename,
 				viteConfig,