+++ packages/integrations/netlify/test/edge-functions/fixtures/edge-basic/astro.config.mjs
@@ -1,6 +1,5 @@
 import { defineConfig } from 'astro/config';
 import { netlifyEdgeFunctions } from '@astrojs/netlify';
-import react from "@astrojs/react";
 
 // test env var
 process.env.SECRET_STUFF = 'secret'
@@ -9,6 +8,5 @@ export default defineConfig({
 	adapter: netlifyEdgeFunctions({
 		dist: new URL('./dist/', import.meta.url),
 	}),
-	integrations: [react()],
 	output: 'server',
 })